configure: Define _DEFAULT_SOURCE in addition to _XOPEN_SOURCE
Defining _XOPEN_SOURCE on its own hides definitions in other
headers (which is already mentioned in commit f49941a89d82ac1329).
One way to deal with this is to define _DEFAULT_SOURCE in addition
to _XOPEN_SOURCE, which is what this commit implements.

This is one way to avoid an implicit function declaration in
the built-in mmap test that autoconf provides.  It includes
<unistd.h> and expects that to provide a declaration of getpagesize
because the earlier check for HAVE_GETPAGESIZE succeeded.  Without
addressing the implicit function declaration, the outcome of this
check will change with future compilers which no longer support
them by default.